fix(storage): fix comments, rename hummock_state_store_metrics
fix(storage): fix grafana
