+++ src/storage/src/hummock/state_store_v1.rs
@@ -103,7 +103,7 @@ impl HummockStorageV1 {
             )
             .await?;
             if let Some(v) = value {
-                local_stats.report(self.state_store_metrics.as_ref(), Some(table_id_label));
+                local_stats.report(self.state_store_metrics.as_ref(), table_id_label);
                 return Ok(v.into_user_value());
             }
             table_counts += table_count;
@@ -118,7 +118,7 @@ impl HummockStorageV1 {
             )
             .await?;
             if let Some(v) = value {
-                local_stats.report(self.state_store_metrics.as_ref(), Some(table_id_label));
+                local_stats.report(self.state_store_metrics.as_ref(), table_id_label);
                 return Ok(v.into_user_value());
             }
             table_counts += table_count;
@@ -151,8 +151,7 @@ impl HummockStorageV1 {
                         )
                         .await?
                         {
-                            local_stats
-                                .report(self.state_store_metrics.as_ref(), Some(table_id_label));
+                            local_stats.report(self.state_store_metrics.as_ref(), table_id_label);
                             return Ok(v.into_user_value());
                         }
                     }
@@ -188,14 +187,14 @@ impl HummockStorageV1 {
                     )
                     .await?
                     {
-                        local_stats.report(self.state_store_metrics.as_ref(), Some(table_id_label));
+                        local_stats.report(self.state_store_metrics.as_ref(), table_id_label);
                         return Ok(v.into_user_value());
                     }
                 }
             }
         }
 
-        local_stats.report(self.state_store_metrics.as_ref(), Some(table_id_label));
+        local_stats.report(self.state_store_metrics.as_ref(), table_id_label);
         self.state_store_metrics
             .iter_merge_sstable_counts
             .with_label_values(&["", "sub-iter"])
@@ -409,7 +408,7 @@ impl HummockStorageV1 {
 
         local_stats.report(
             self.state_store_metrics.as_ref(),
-            Some(table_id.to_string().as_str()),
+            table_id.to_string().as_str(),
         );
         Ok(HummockStateStoreIter::new(
             user_iterator,
@@ -636,6 +635,6 @@ impl Drop for HummockStateStoreIter {
     fn drop(&mut self) {
         let mut stats = StoreLocalStatistic::default();
         self.collect_local_statistic(&mut stats);
-        stats.report(&self.metrics, None);
+        stats.report(&self.metrics, "");
     }
 }
